refactor(staging): extract file-based store to dedicated package
Move file-based staging storage implementation to `internal/staging/file`
package, parallel to future `agent` package structure.

Breaking changes:
- `staging.Store` → `file.Store`
- `staging.NewStore` → `file.NewStore`
- `staging.NewStoreWithPath` → `file.NewStoreWithPath`

The `staging` package now only contains shared types:
- Entry, TagEntry, State, Service, Operation
- ErrNotStaged error
- StoreReader, StoreWriter, StoreReadWriter interfaces
