+++ internal/staging/runner/add_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/mpyw/suve/internal/staging"
+	"github.com/mpyw/suve/internal/staging/file"
 	"github.com/mpyw/suve/internal/staging/runner"
 	stagingusecase "github.com/mpyw/suve/internal/usecase/staging"
 )
@@ -23,7 +24,7 @@ func TestAddRunner_Run(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var buf bytes.Buffer
 		r := &runner.AddRunner{
@@ -52,7 +53,7 @@ func TestAddRunner_Run(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		// Pre-stage as create
 		_ = store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
@@ -89,7 +90,7 @@ func TestAddRunner_Run(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var buf bytes.Buffer
 		r := &runner.AddRunner{
@@ -115,7 +116,7 @@ func TestAddRunner_Run(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		// Pre-stage as create
 		_ = store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
@@ -145,7 +146,7 @@ func TestAddRunner_Run(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var buf bytes.Buffer
 		r := &runner.AddRunner{
@@ -201,7 +202,7 @@ func TestAddRunner_ErrorCases(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var stdout, stderr bytes.Buffer
 		r := &runner.AddRunner{
@@ -222,7 +223,7 @@ func TestAddRunner_ErrorCases(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var stdout, stderr bytes.Buffer
 		r := &runner.AddRunner{
@@ -250,7 +251,7 @@ func TestAddRunner_WithOptions(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var stdout, stderr bytes.Buffer
 		r := &runner.AddRunner{
@@ -280,7 +281,7 @@ func TestAddRunner_WithOptions(t *testing.T) {
 		t.Parallel()
 
 		tmpDir := t.TempDir()
-		store := staging.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
+		store := file.NewStoreWithPath(filepath.Join(tmpDir, "stage.json"))
 
 		var stdout, stderr bytes.Buffer
 		r := &runner.AddRunner{
