+++ internal/usecase/staging/add_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/mpyw/suve/internal/staging"
+	"github.com/mpyw/suve/internal/staging/file"
 	"github.com/mpyw/suve/internal/staging/transition"
 	usecasestaging "github.com/mpyw/suve/internal/usecase/staging"
 )
@@ -44,7 +45,7 @@ func newMockParser() *mockParser {
 func TestAddUseCase_Execute(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.AddUseCase{
 		Strategy: newMockEditStrategyNotFound(), // Resource doesn't exist - expected for add
 		Store:    store,
@@ -69,7 +70,7 @@ func TestAddUseCase_Execute(t *testing.T) {
 func TestAddUseCase_Execute_RejectsWhenResourceExists(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	// Resource exists on AWS
 	uc := &usecasestaging.AddUseCase{
 		Strategy: newMockEditStrategy(), // Returns existing value
@@ -87,7 +88,7 @@ func TestAddUseCase_Execute_RejectsWhenResourceExists(t *testing.T) {
 func TestAddUseCase_Execute_MinimalInput(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.AddUseCase{
 		Strategy: newMockEditStrategyNotFound(),
 		Store:    store,
@@ -108,7 +109,7 @@ func TestAddUseCase_Execute_MinimalInput(t *testing.T) {
 func TestAddUseCase_Draft_NotStaged(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.AddUseCase{
 		Strategy: newMockEditStrategyNotFound(),
 		Store:    store,
@@ -123,7 +124,7 @@ func TestAddUseCase_Draft_NotStaged(t *testing.T) {
 func TestAddUseCase_Draft_StagedCreate(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/draft", staging.Entry{
 		Operation: staging.OperationCreate,
 		Value:     lo.ToPtr("draft-value"),
@@ -144,7 +145,7 @@ func TestAddUseCase_Draft_StagedCreate(t *testing.T) {
 func TestAddUseCase_Draft_StagedUpdate(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	// Update operation should not be returned as draft
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/update", staging.Entry{
 		Operation: staging.OperationUpdate,
@@ -165,7 +166,7 @@ func TestAddUseCase_Draft_StagedUpdate(t *testing.T) {
 func TestAddUseCase_Execute_ParseError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockEditStrategyNotFound()
 	strategy.parseErr = errors.New("invalid name")
 
@@ -185,7 +186,7 @@ func TestAddUseCase_Execute_ParseError(t *testing.T) {
 func TestAddUseCase_Draft_ParseError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockEditStrategyNotFound()
 	strategy.parseErr = errors.New("invalid name")
 
@@ -255,7 +256,7 @@ func TestAddUseCase_Execute_GetError(t *testing.T) {
 func TestAddUseCase_Execute_RejectsWhenUpdateStaged(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	// Pre-stage an UPDATE operation
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/existing", staging.Entry{
 		Operation: staging.OperationUpdate,
@@ -279,7 +280,7 @@ func TestAddUseCase_Execute_RejectsWhenUpdateStaged(t *testing.T) {
 func TestAddUseCase_Execute_RejectsWhenDeleteStaged(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	// Pre-stage a DELETE operation
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/deleted", staging.Entry{
 		Operation: staging.OperationDelete,
@@ -302,7 +303,7 @@ func TestAddUseCase_Execute_RejectsWhenDeleteStaged(t *testing.T) {
 func TestAddUseCase_Execute_AllowsReEditOfCreate(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	// Pre-stage a CREATE operation
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/new", staging.Entry{
 		Operation: staging.OperationCreate,
@@ -333,7 +334,7 @@ func TestAddUseCase_Execute_AllowsReEditOfCreate(t *testing.T) {
 func TestAddUseCase_Execute_FetchError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockEditStrategy()
 	strategy.fetchErr = errors.New("AWS connection error")
