+++ internal/usecase/staging/apply_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/mpyw/suve/internal/staging"
+	"github.com/mpyw/suve/internal/staging/file"
 	usecasestaging "github.com/mpyw/suve/internal/usecase/staging"
 )
 
@@ -54,7 +55,7 @@ func newMockApplyStrategy() *mockApplyStrategy {
 func TestApplyUseCase_Execute_Empty(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.ApplyUseCase{
 		Strategy: newMockApplyStrategy(),
 		Store:    store,
@@ -72,7 +73,7 @@ func TestApplyUseCase_Execute_Empty(t *testing.T) {
 func TestApplyUseCase_Execute_SingleCreate(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/new", staging.Entry{
 		Operation: staging.OperationCreate,
 		Value:     lo.ToPtr("new-value"),
@@ -101,7 +102,7 @@ func TestApplyUseCase_Execute_SingleCreate(t *testing.T) {
 func TestApplyUseCase_Execute_MultipleOperations(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/create", staging.Entry{
 		Operation: staging.OperationCreate,
 		Value:     lo.ToPtr("create"),
@@ -133,7 +134,7 @@ func TestApplyUseCase_Execute_MultipleOperations(t *testing.T) {
 func TestApplyUseCase_Execute_FilterByName(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/one", staging.Entry{
 		Operation: staging.OperationUpdate,
 		Value:     lo.ToPtr("one"),
@@ -167,7 +168,7 @@ func TestApplyUseCase_Execute_FilterByName(t *testing.T) {
 func TestApplyUseCase_Execute_FilterByName_NotStaged(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.ApplyUseCase{
 		Strategy: newMockApplyStrategy(),
 		Store:    store,
@@ -183,7 +184,7 @@ func TestApplyUseCase_Execute_FilterByName_NotStaged(t *testing.T) {
 func TestApplyUseCase_Execute_PartialFailure(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/success", staging.Entry{
 		Operation: staging.OperationUpdate,
 		Value:     lo.ToPtr("success"),
@@ -226,7 +227,7 @@ func TestApplyUseCase_Execute_ConflictDetection(t *testing.T) {
 	baseTime := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
 	awsTime := time.Date(2024, 1, 2, 0, 0, 0, 0, time.UTC) // Modified after staging
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/conflict", staging.Entry{
 		Operation:      staging.OperationUpdate,
 		Value:          lo.ToPtr("staged"),
@@ -270,7 +271,7 @@ func TestApplyUseCase_Execute_ListError(t *testing.T) {
 func TestApplyUseCase_Execute_DeleteSuccess(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/to-delete", staging.Entry{
 		Operation: staging.OperationDelete,
 		StagedAt:  time.Now(),
@@ -316,7 +317,7 @@ func newMockApplyTagStrategy() *mockApplyTagStrategy {
 func TestApplyUseCase_Execute_TagsOnly(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
 		Add:      map[string]string{"env": "prod", "team": "backend"},
 		StagedAt: time.Now(),
@@ -345,7 +346,7 @@ func TestApplyUseCase_Execute_TagsOnly(t *testing.T) {
 func TestApplyUseCase_Execute_TagsWithRemove(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	removeKeys := map[string]struct{}{"deprecated": {}, "old": {}}
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
 		Add:      map[string]string{"env": "prod"},
@@ -369,7 +370,7 @@ func TestApplyUseCase_Execute_TagsWithRemove(t *testing.T) {
 func TestApplyUseCase_Execute_EntriesAndTags(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
 		Operation: staging.OperationUpdate,
 		Value:     lo.ToPtr("new-value"),
@@ -396,7 +397,7 @@ func TestApplyUseCase_Execute_EntriesAndTags(t *testing.T) {
 func TestApplyUseCase_Execute_TagFailure(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
 		Add:      map[string]string{"env": "prod"},
 		StagedAt: time.Now(),
@@ -426,7 +427,7 @@ func TestApplyUseCase_Execute_TagFailure(t *testing.T) {
 func TestApplyUseCase_Execute_PartialTagFailure(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/success", staging.TagEntry{
 		Add:      map[string]string{"env": "prod"},
 		StagedAt: time.Now(),
@@ -462,7 +463,7 @@ func TestApplyUseCase_Execute_PartialTagFailure(t *testing.T) {
 func TestApplyUseCase_Execute_FilterByName_TagOnly(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/one", staging.TagEntry{
 		Add:      map[string]string{"env": "prod"},
 		StagedAt: time.Now(),
