+++ internal/usecase/staging/reset_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/mpyw/suve/internal/staging"
+	"github.com/mpyw/suve/internal/staging/file"
 	usecasestaging "github.com/mpyw/suve/internal/usecase/staging"
 )
 
@@ -50,7 +51,7 @@ func newMockResetStrategy() *mockResetStrategy {
 func TestResetUseCase_Execute_Unstage(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
 		Operation: staging.OperationUpdate,
 		Value:     lo.ToPtr("value"),
@@ -77,7 +78,7 @@ func TestResetUseCase_Execute_Unstage(t *testing.T) {
 func TestResetUseCase_Execute_NotStaged(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.ResetUseCase{
 		Parser: newMockParser(),
 		Store:  store,
@@ -93,7 +94,7 @@ func TestResetUseCase_Execute_NotStaged(t *testing.T) {
 func TestResetUseCase_Execute_UnstageAll(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/one", staging.Entry{
 		Operation: staging.OperationUpdate,
 		Value:     lo.ToPtr("one"),
@@ -126,7 +127,7 @@ func TestResetUseCase_Execute_UnstageAll(t *testing.T) {
 func TestResetUseCase_Execute_UnstageAll_Empty(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.ResetUseCase{
 		Parser: newMockParser(),
 		Store:  store,
@@ -142,7 +143,7 @@ func TestResetUseCase_Execute_UnstageAll_Empty(t *testing.T) {
 func TestResetUseCase_Execute_Restore(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
@@ -171,7 +172,7 @@ func TestResetUseCase_Execute_Restore(t *testing.T) {
 func TestResetUseCase_Execute_Restore_NoFetcher(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
@@ -193,7 +194,7 @@ func TestResetUseCase_Execute_Restore_NoFetcher(t *testing.T) {
 func TestResetUseCase_Execute_Restore_FetchError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
@@ -226,7 +227,7 @@ func (m *mockParserWithVersion) ParseSpec(input string) (string, bool, error) {
 func TestResetUseCase_Execute_ParseError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithParseSpecErr{
 		mockParser: newMockParser(),
 		parseErr:   errors.New("parse error"),
@@ -347,7 +348,7 @@ func TestResetUseCase_Execute_RestoreStageError(t *testing.T) {
 func TestResetUseCase_Execute_RestoreSkipped_SameAsAWS(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
@@ -380,7 +381,7 @@ func TestResetUseCase_Execute_RestoreSkipped_SameAsAWS(t *testing.T) {
 func TestResetUseCase_Execute_RestoreNotSkipped_DifferentFromAWS(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
@@ -414,7 +415,7 @@ func TestResetUseCase_Execute_RestoreNotSkipped_DifferentFromAWS(t *testing.T) {
 func TestResetUseCase_Execute_RestoreFetchCurrentError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	parser := &mockParserWithVersion{
 		mockParser: newMockParser(),
 		hasVersion: true,
