+++ internal/usecase/staging/tag_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/mpyw/suve/internal/maputil"
 	"github.com/mpyw/suve/internal/staging"
+	"github.com/mpyw/suve/internal/staging/file"
 	usecasestaging "github.com/mpyw/suve/internal/usecase/staging"
 )
 
@@ -46,7 +47,7 @@ func newMockTagStrategy() *mockTagStrategy {
 func TestTagUseCase_Tag_NewTagEntry(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.TagUseCase{
 		Strategy: newMockTagStrategy(),
 		Store:    store,
@@ -70,7 +71,7 @@ func TestTagUseCase_Tag_NewTagEntry(t *testing.T) {
 func TestTagUseCase_Tag_MergeWithExisting(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	baseTime := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
 
 	// Pre-stage a tag entry
@@ -102,7 +103,7 @@ func TestTagUseCase_Tag_MergeWithExisting(t *testing.T) {
 func TestTagUseCase_Tag_AddTagRemovesFromUntagList(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Pre-stage with remove tags
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
@@ -132,7 +133,7 @@ func TestTagUseCase_Tag_AddTagRemovesFromUntagList(t *testing.T) {
 func TestTagUseCase_Tag_ParseError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.parseErr = errors.New("invalid name")
 
@@ -152,7 +153,7 @@ func TestTagUseCase_Tag_ParseError(t *testing.T) {
 func TestTagUseCase_Tag_FetchError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.fetchErr = errors.New("aws error")
 
@@ -210,7 +211,7 @@ func TestTagUseCase_Tag_GetError(t *testing.T) {
 func TestTagUseCase_Tag_ZeroLastModified(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.fetchResult = &staging.EditFetchResult{
 		Value:        "aws-value",
@@ -236,7 +237,7 @@ func TestTagUseCase_Tag_ZeroLastModified(t *testing.T) {
 func TestTagUseCase_Tag_StagedForCreate_ResourceNotFound(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// First, stage an entry for CREATE (new resource that doesn't exist in AWS)
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/new-param", staging.Entry{
@@ -271,7 +272,7 @@ func TestTagUseCase_Tag_StagedForCreate_ResourceNotFound(t *testing.T) {
 func TestTagUseCase_Tag_EmptyTags(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.TagUseCase{
 		Strategy: newMockTagStrategy(),
 		Store:    store,
@@ -289,7 +290,7 @@ func TestTagUseCase_Tag_EmptyTags(t *testing.T) {
 func TestTagUseCase_Tag_BlockedOnDelete(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Pre-stage DELETE operation
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
@@ -337,7 +338,7 @@ func TestTagUseCase_Tag_GetEntryError(t *testing.T) {
 func TestTagUseCase_Untag_Success(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.TagUseCase{
 		Strategy: newMockTagStrategy(),
 		Store:    store,
@@ -360,7 +361,7 @@ func TestTagUseCase_Untag_Success(t *testing.T) {
 func TestTagUseCase_Untag_RemoveTagDeletesFromAddList(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Pre-stage with tags to add
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
@@ -390,7 +391,7 @@ func TestTagUseCase_Untag_RemoveTagDeletesFromAddList(t *testing.T) {
 func TestTagUseCase_Untag_DuplicateRemoveKeys(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Pre-stage with remove key
 	require.NoError(t, store.StageTag(t.Context(), staging.ServiceParam, "/app/config", staging.TagEntry{
@@ -420,7 +421,7 @@ func TestTagUseCase_Untag_DuplicateRemoveKeys(t *testing.T) {
 func TestTagUseCase_Untag_ParseError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.parseErr = errors.New("invalid name")
 
@@ -478,7 +479,7 @@ func TestTagUseCase_Untag_GetTagError(t *testing.T) {
 func TestTagUseCase_Untag_FetchError(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.fetchErr = errors.New("aws error")
 
@@ -498,7 +499,7 @@ func TestTagUseCase_Untag_FetchError(t *testing.T) {
 func TestTagUseCase_Untag_BlockedOnDelete(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Pre-stage DELETE operation
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/config", staging.Entry{
@@ -523,7 +524,7 @@ func TestTagUseCase_Untag_BlockedOnDelete(t *testing.T) {
 func TestTagUseCase_Untag_StagedForCreate_WithExistingTags(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Stage an entry for CREATE
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/new-param", staging.Entry{
@@ -566,7 +567,7 @@ func TestTagUseCase_Untag_StagedForCreate_WithExistingTags(t *testing.T) {
 func TestTagUseCase_Untag_StagedForCreate_AutoSkip(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 
 	// Stage an entry for CREATE without any pre-staged tags
 	require.NoError(t, store.StageEntry(t.Context(), staging.ServiceParam, "/app/new-param", staging.Entry{
@@ -599,7 +600,7 @@ func TestTagUseCase_Untag_StagedForCreate_AutoSkip(t *testing.T) {
 func TestTagUseCase_Untag_EmptyTagKeys(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	uc := &usecasestaging.TagUseCase{
 		Strategy: newMockTagStrategy(),
 		Store:    store,
@@ -636,7 +637,7 @@ func TestTagUseCase_Untag_StageError(t *testing.T) {
 func TestTagUseCase_Untag_PreservesBaseModifiedAt(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	baseTime := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
 
 	// Pre-stage a tag entry with BaseModifiedAt
@@ -665,7 +666,7 @@ func TestTagUseCase_Untag_PreservesBaseModifiedAt(t *testing.T) {
 func TestTagUseCase_Untag_ZeroLastModified(t *testing.T) {
 	t.Parallel()
 
-	store := staging.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
+	store := file.NewStoreWithPath(filepath.Join(t.TempDir(), "staging.json"))
 	strategy := newMockTagStrategy()
 	strategy.fetchResult = &staging.EditFetchResult{
 		Value:        "aws-value",
